Implement copy_after and copy_before and fix bugs
Multiple prerequisites are now correctly handled.

Copy, which default to copy_after, was accompanied with
copy_before.

Links to rule names in prerequisites are allowed, though
do not work until the flag `--flags-only` is implemented.

The case when target which uses global variable was not
formatted propertly is now fixed.